[DEV-5733] Move calc_metrics_prs tests to a separate file
